backend/requirements.txt
django
django_rest_framework
xlrd
requests
psycopg2
django-cors-headers